--- a/x-pack/plugins/ingest_manager/server/services/agents/actions.ts
+++ b/x-pack/plugins/ingest_manager/server/services/agents/actions.ts
@@ -27,22 +27,21 @@ export async function createAgentAction(
 
 export async function getAgentActionsForCheckin(
   soClient: SavedObjectsClientContract,
-  agentId: string
+  agent: Agent
 ): Promise<AgentAction[]> {
-  const res = await soClient.find<AgentActionSOAttributes>({
-    type: AGENT_ACTION_SAVED_OBJECT_TYPE,
-    filter: `not ${AGENT_ACTION_SAVED_OBJECT_TYPE}.attributes.sent_at: * and ${AGENT_ACTION_SAVED_OBJECT_TYPE}.attributes.agent_id:${agentId}`,
-  });
+  if (agent.not_acknowledged_actions.length === 0) {
+    return [];
+  }
 
   return Promise.all(
-    res.saved_objects.map(async (so) => {
+    agent.not_acknowledged_actions.map(async (actionId) => {
       // Get decrypted actions
       return savedObjectToAgentAction(
         await appContextService
           .getEncryptedSavedObjects()
           .getDecryptedAsInternalUser<AgentActionSOAttributes>(
             AGENT_ACTION_SAVED_OBJECT_TYPE,
-            so.id
+            actionId
           )
       );
     })
@@ -80,7 +79,7 @@ export async function getAgentActionByIds(
 export async function getNewActionsSince(soClient: SavedObjectsClientContract, timestamp: string) {
   const res = await soClient.find<AgentActionSOAttributes>({
     type: AGENT_ACTION_SAVED_OBJECT_TYPE,
-    filter: `not ${AGENT_ACTION_SAVED_OBJECT_TYPE}.attributes.sent_at: * AND ${AGENT_ACTION_SAVED_OBJECT_TYPE}.attributes.created_at >= "${timestamp}"`,
+    filter: `not ${AGENT_ACTION_SAVED_OBJECT_TYPE}.attributes.acknowledged_at: * AND ${AGENT_ACTION_SAVED_OBJECT_TYPE}.attributes.created_at >= "${timestamp}"`,
   });
 
   return res.saved_objects.map(savedObjectToAgentAction);