--- a/x-pack/plugins/ingest_manager/server/services/agents/checkin/index.ts
+++ b/x-pack/plugins/ingest_manager/server/services/agents/checkin/index.ts
@@ -44,7 +44,7 @@ export async function agentCheckin(
   }
 
   // Check if some actions are not acknowledged
-  let actions = await getAgentActionsForCheckin(soClient, agent.id);
+  let actions = await getAgentActionsForCheckin(soClient, agent);
   if (actions.length > 0) {
     return { actions };
   }