--- a/x-pack/plugins/ingest_manager/server/types/models/agent.ts
+++ b/x-pack/plugins/ingest_manager/server/types/models/agent.ts
@@ -69,5 +69,4 @@ export const NewAgentActionSchema = schema.object({
     schema.literal('PAUSE'),
   ]),
   data: schema.maybe(schema.any()),
-  sent_at: schema.maybe(schema.string()),
 });