@jen-huang 2020-08-25
How will this affect agent actions created in 7.9? I see saved object mapping changes. Does there need to be migrations added if there are documents from 7.9?