@nchaulet 2020-08-25
@jen-huang it will affect actions from 7.9, I am planning on writing the migration, but we should have an happy path to migrate  the renaming of `sent_at` to `acknowledged_at`